Move range validation to AbstractTabletFile
Also add validation to StoredTabletFile validation method and on
serialization
