+++ core/src/main/java/org/apache/accumulo/core/metadata/ReferencedTabletFile.java
@@ -19,7 +19,6 @@
 package org.apache.accumulo.core.metadata;
 
 import static org.apache.accumulo.core.Constants.HDFS_TABLES_DIR;
-import static org.apache.accumulo.core.metadata.StoredTabletFile.requireRowRange;
 
 import java.net.URI;
 import java.util.Comparator;
@@ -166,7 +165,7 @@ public class ReferencedTabletFile extends AbstractTabletFile<ReferencedTabletFil
    * qualify an absolute path or create a new file.
    */
   public ReferencedTabletFile(Path metaPath, Range range) {
-    super(Objects.requireNonNull(metaPath), requireRowRange(range));
+    super(Objects.requireNonNull(metaPath), range);
     log.trace("Parsing TabletFile from {}", metaPath);
     parts = parsePath(metaPath);
   }
